[region-isolation] Add a test showing that we do not infer nonisolated(unsafe) onto named closures.
The semantics we want is for nonisolated(unsafe) to be inferred onto anonymous
closures. We already have that semantics since we do not look through move_value
[var_decl] or begin_borrow [var_decl]. So even though the underlying
partial_apply will be viewed as nonisolated(unsafe), we will not consider that
when determining the nonisolated(unsafe) for the begin_borrow or move_value that
defined the variable whose value is the partial_apply.

(cherry picked from commit 359eeec7466b11998a89490c75994bc1cf70af9c)