Update CI summary table on README to be more compact (#81647)
This PR updates the CI summary table in the README.md to summarize all
architectures for a host to be on the same line with the architecture
specified in the badge itself (rather than a generic "build" label),
thereby enabling the whole CI status to be visible on a single page.

It also adds accessibility labels to each of the CI badges for Global
Accessibility Awareness Month.