build: Skip configure when no variable has changed
`build.ps1` always re-runs the configuration step before building since
some of the CMake variables might have changed between 2 runs. This
takes time in incremental builds for something that is not required most
of the time.

These changes read `CMakeCache.txt` and compare the values in the file
with the ones that would be used in the configure invocation and skips
it if they are identical. This greatly speeds up incremental builds.

Local testing of a no-op incremental build has shown an improvement of
almost 3 minutes in build time on Windows x64 building all 3 SDKs with
no packaging, going from 6min10s before these changes to 3min30s with
these changes.