Don't disable priority escalation tests on arm64e
We carried over all unsupported: marks from task_priority test which seems to have had just racy behavior for ages. This test should be fine on arm64e, and especially, we must run it to verify the pointer auth of these APIs.

This allowed a pointer auth issue to slip through, so let's remedy this by actually running the test on those platforms.