Don't expect PCMPEQQ on x86_64
This instruction was added in SSE4.1, which is in the baseline that macOS targets now, but not in the baseline for older macOS targets, nor for Linux or Windows. Eventually, we'll want to be able to slice these tests more finely by ISA extension targeted, but the arm64 coverage gets us what we need here.