LoopRotate: Fix a by reference map bug under reallocation
When using a densemap subscript expression on both sides of an
assignment in the same statement of the same map we run into the issue
that the map can reallocate because of the assignment but we are
referencing the value of the RHS map subscript by reference --
i.e we can reference deallocated memory.

Not good.

My attempts at making a reproducer crash failed.

rdar://151031297